// src/app/views/dashboard/Setup/tuition/models/tuitions.ts
export class Tuitions {
  id?: number;
  dni?: string;
  apellidos?: string;
  nombres?: string;
  especialidad?: string;
  modulo?: string;
  turno?: string;
  hora_inicio?: string;
  hora_fin?: string;
  voucher_img?: string;
  modalidad?: string;
  seccion?: string;
  observacion?: string;
  fecha?: string;
  pago?: string;
  codigo?: string;
  voucher_aprobado?: boolean;
}